requirements.txt
# Core
fastapi
uvicorn
pydantic
httpx
httpx[http2]
xmltodict
asyncio
python-dotenv

# Dependencies for ddtrace compatibility
opentelemetry-api
protobuf
wrapt

# Speed-ups (Linux/macOS only)
uvloop; platform_system != "Windows"
httptools; platform_system != "Windows"

# Dev/test
pytest
pytest-asyncio
